fix: userid 호출방식 변경

+++ src/main/java/org/scoula/service/learning/LearningService.java
@@ -16,8 +16,7 @@ public interface LearningService {
     int getUserCredit(int userId);
     boolean checkQuiz(int userId, int quizId);
     boolean hasCompleted(int userId, int contentId);
-    //List<LearningHistoryDto> getLearningHistoryList(int userId);
     void saveResult(int userId, int quizId, boolean isCorrect, String selectedAnswer, int creditEarned);
     QuizResultDTO getQuizResult(int userId, int quizId);
-    List<LearningContentDTO> getCompletedContents(Long userId);
+    List<LearningContentDTO> getCompletedContents(int userId);
 }
